utils: Move `NextDown`
